Move ExtensionCandidate to ergo-chain-types, fix NipopowProof proptests
